Extract useDependsOn hook and use in selectors
Introduce a new useDependsOn hook to centralize tracking of dependent parameter values and validity. Refactor Choice.jsx and Slider.jsx to consume the hook (removing duplicated useMemo logic) so selectors only fetch data when their dependent parameters are set. Slider.jsx also skips dynamic fetch when a static range is provided, waits for dependsValid before fetching, and includes dependsOnValues/dependsValid in its effect deps. Update base.jsx to pass parameter.depends_on into the Slider as dependsOn. These changes reduce duplication and prevent premature/duplicate fetches tied to parameter dependencies.
